+++ CHANGELOG.md
@@ -7,9 +7,13 @@ All notable changes to this project will be documented in this file.
 ### Added
 - [FRONTEND] Add new background and banner images for innovations
 
+- [FRONTEND] Add solution partner cards
 - [BACKEND] Add 5 new seed data for partner solution service
 - [BACKEND] Add new service type: `solution-partner`
 - [BACKEND] Add new controller to get services with service type: `solution-partner`
+- [BACKEND] Add new predefined visitor in seed that will be used for a solution partner
+- [BACKEND] Update visitor activities endpoint for solution partner
+- [BACKEND] Take out the unique index in service's slug
 
 ### Changed